Move field node source methods to trait
