+++ docroot/modules/custom/sitenow_migrate/src/Plugin/migrate/source/BaseNodeSource.php
@@ -24,6 +24,7 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
  * @see \Drupal\node\Plugin\migrate\source\d7\Node
  */
 abstract class BaseNodeSource extends SqlBase {
+  use FieldableEntityTrait;
 
   /**
    * The module handler.
@@ -301,77 +302,6 @@ abstract class BaseNodeSource extends SqlBase {
     // Run garbage collector to further reduce memory.
     gc_collect_cycles();
     return memory_get_usage();
-  }
-
-  /**
-   * Returns all non-deleted field instances attached to a specific entity type.
-   *
-   * @param string $entity_type
-   *   The entity type ID.
-   * @param string|null $bundle
-   *   (optional) The bundle.
-   *
-   * @return array[]
-   *   The field instances, keyed by field name.
-   */
-  protected function getFields($entity_type, $bundle = NULL) {
-    $query = $this->select('field_config_instance', 'fci')
-      ->fields('fci')
-      ->condition('fci.entity_type', $entity_type)
-      ->condition('fci.bundle', isset($bundle) ? $bundle : $entity_type)
-      ->condition('fci.deleted', 0);
-
-    // Join the 'field_config' table and add the 'translatable' setting to the
-    // query.
-    $query->leftJoin('field_config', 'fc', 'fci.field_id = fc.id');
-    $query->addField('fc', 'translatable');
-
-    return $query->execute()->fetchAllAssoc('field_name');
-  }
-
-  /**
-   * Retrieves field values for a single field of a single entity.
-   *
-   * @param string $entity_type
-   *   The entity type.
-   * @param string $field
-   *   The field name.
-   * @param int $entity_id
-   *   The entity ID.
-   * @param int|null $revision_id
-   *   (optional) The entity revision ID.
-   * @param string $language
-   *   (optional) The field language.
-   *
-   * @return array
-   *   The raw field values, keyed by delta.
-   */
-  protected function getFieldValues($entity_type, $field, $entity_id, $revision_id = NULL, $language = NULL) {
-    $table = (isset($revision_id) ? 'field_revision_' : 'field_data_') . $field;
-    $query = $this->select($table, 't')
-      ->fields('t')
-      ->condition('entity_type', $entity_type)
-      ->condition('entity_id', $entity_id)
-      ->condition('deleted', 0);
-    if (isset($revision_id)) {
-      $query->condition('revision_id', $revision_id);
-    }
-    // Add 'language' as a query condition if it has been defined by Entity
-    // Translation.
-    if ($language) {
-      $query->condition('language', $language);
-    }
-    $values = [];
-    foreach ($query->execute() as $row) {
-      foreach ($row as $key => $value) {
-        $delta = $row['delta'];
-        if (strpos($key, $field) === 0) {
-          $column = substr($key, strlen($field) + 1);
-          $values[$delta][$column] = $value;
-        }
-      }
-    }
-    return $values;
   }
 
 }
